requirements.txt
streamlit
google-generativeai
huggingface_hub
diffusers
torch
pillow
python-dotenv
